centralize spatial core a bit
